+++ mod.ts
@@ -6,4 +6,5 @@ export * from "./src/HtmlShell.tsx";
 export * from "./src/ssr.ts";
 export * from "./src/types.ts";
 export * from "./src/useFullSoak.ts";
-export { setupDefaultFullsoakLogger } from "./src/utils.ts";
+export { getOrigin } from "./src/metastore.ts";
+export { getComponentJs, setupDefaultFullsoakLogger } from "./src/utils.ts";