+++ src/CsrController.ts
@@ -4,11 +4,61 @@ import { getClientSideJsForRoute } from "./getClientSideJsForRoute.ts";
 import * as Uglify from "uglify-js";
 import { cleanCss } from "./minifyCss.ts";
 import { getGlobalComponentsDir } from "./metastore.ts";
-import { type Output, transformFile } from "@swc/core";
-import { getComponentCss, IS_DEBUG, LogDebug, LogError } from "./utils.ts";
+import { bundle, type Output, transform } from "@swc/core";
+import {
+  DENO_DIR,
+  getComponentCss,
+  getComponentJs,
+  LogError,
+} from "./utils.ts";
 
 @Controller()
 export class CsrController {
+  @Get("/fullsoak")
+  async serveFullsoak(ctx: Context): Promise<string> {
+    ctx.response.headers.set("content-type", "text/javascript");
+    // @NOTE we can export other things as well; this is super helpful
+    // for isomorphic code imported via the same handler on both
+    // server and client sides, each requiring different a implementation
+    const res = await transform(
+      `export const getOrigin = () => window.location.origin;`,
+    );
+    return res.code;
+  }
+
+  @Get("/preact-iso")
+  async servePreactIso(ctx: Context): Promise<string> {
+    ctx.response.headers.set("content-type", "text/javascript");
+
+    const NPM_DIR = `${DENO_DIR}/deno/npm/registry.npmjs.org`;
+    // @TODO take care of preact-iso versioning below
+    const preactIso = `${NPM_DIR}/preact-iso/2.8.1/src/index.js`;
+
+    let output: Record<string, Output> = {};
+    try {
+      output = await bundle({
+        entry: preactIso,
+        target: "browser",
+        externalModules: ["preact", "preact/hooks"],
+        // deno-lint-ignore no-explicit-any
+      } as any);
+    } catch (e) {
+      LogError(
+        "failed to bundle preact-iso",
+        JSON.stringify((e as Error).stack),
+      );
+    }
+    return output["index.js"]?.code ||
+      "throw new Error('failed to serve preact-iso')";
+    // const mod = await import("preact-iso");
+    // let retVal = "";
+    // for (const key of Object.keys(mod)) {
+    //   const func = mod[key as keyof typeof import("preact-iso")];
+    //   retVal += `export const ${key} = ${func};`;
+    // }
+    // return retVal;
+  }
+
   @Get("/js/:compName/mount.js")
   @ControllerMethodArgs("param")
   serveClientJsEntryPoint(
@@ -47,46 +97,7 @@ export class CsrController {
 
     const compFile = `${compName}/index.tsx`;
     const fullCompFile = `${globalComponentsDir}/${compFile}`;
-    LogDebug("transforming component", fullCompFile);
 
-    let transformedComp: Output;
-
-    try {
-      transformedComp = await transformFile(fullCompFile, {
-        env: {
-          debug: IS_DEBUG && true,
-        },
-        jsc: {
-          parser: {
-            syntax: "typescript",
-            tsx: true,
-          },
-          transform: {
-            react: {
-              runtime: "automatic",
-              pragma: "h",
-              pragmaFrag: "Fragment",
-              refresh: true, // @TODO disable for prod
-            },
-            // "optimizer": {
-            //   "globals": {
-            //     "vars": {
-            //       "__DEBUG__": "true",
-            //     },
-            //   },
-            // },
-          },
-        },
-      });
-    } catch (e) {
-      LogError("error transforming component", {
-        path: fullCompFile,
-        error: e,
-      });
-      return `console.error("error loading component" '${compFile}');`;
-    }
-
-    // @TODO consider what to do with source map
-    return transformedComp.code;
+    return await getComponentJs(fullCompFile);
   }
 }